progress towards nginx+uwsgi
Running into a lot of pynag / adagios bugs, and I'm not happy about
it. Forked both projects and pointing pip at github.

The uwsgi part is relatively straightforward; some of the salt states
are kinda messy and probably just need to be combined into one file;
Unlike shinken or graphite it doesn't make any sense to install adagios
without a web frontend.

The only tricky part of wsgi is permissions; nginx runs as www-data, but
adagios needs to run as shinken so it has permission to mess with
shinken files and daemons.

We are still monkeypatching pynag to work against shinken better;
There's a game of whack-a-mole there, but I'm working through errors one
by one.

refs DataCenter:#389 (4)